requirements.txt
Flask~=3.0
waitress~=3.0
validators~=0.0
pillow~=10.0